BASWSPRT-308: Refactor custom field filter logic into a lone class.
